style: refactoring in index.ts and tackling todos
